Keep test fixtures importable
We skip *_test.py files when building wheels (or doing non-editable
installs in general), but one such file was referenced by an
"importable" fixture
